+++ android/app-newm/src/main/java/io/newm/HomeActivity.kt
@@ -1,18 +1,13 @@
 package io.newm
 
-import android.app.Activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowCompat
 import com.slack.circuit.foundation.Circuit
 import com.slack.circuit.foundation.CircuitCompositionLocals
 import com.slack.circuit.retained.LocalRetainedStateRegistry
@@ -48,8 +43,8 @@ import io.newm.screens.studio.StudioPresenter
 import io.newm.screens.studio.StudioScreenUi
 import io.newm.screens.studio.StudioState
 import io.newm.screens.wallets.WalletsPresenter
-import io.newm.screens.wallets.view.WalletsUi
 import io.newm.screens.wallets.WalletsUiState
+import io.newm.screens.wallets.view.WalletsUi
 import io.newm.shared.NewmAppLogger
 import io.newm.shared.public.analytics.NewmAppEventLogger
 import io.newm.shared.public.analytics.events.AppScreens
