--- android/app-newm/src/main/java/io/newm/screens/home/MockData.kt
@@ -1,119 +1,0 @@
-package io.newm.screens.home
-
-val exploreMusicModels = listOf(
-    MusicModel(
-        title = "Indie music",
-        subtitle = "Playlist",
-        imageUrl = "https://www.hollywoodinsider.com/wp-content/uploads/2021/03/Hollywood-Insider-Indie-and-DIY-Music.jpg",
-        imageText = "Indie"
-    ),
-    MusicModel(
-        title = "Country music",
-        subtitle = "Playlist",
-        imageUrl = "https://www.dictionary.com/e/wp-content/uploads/2019/04/1000x700-country-music-v1.jpg",
-        imageText = "Country"
-    ),
-    MusicModel(
-        title = "Soft music",
-        subtitle = "Playlist",
-        imageUrl = "https://i.ytimg.com/vi/kqA5Y_tpbAw/maxresdefault.jpg",
-        imageText = "Soft"
-    )
-)
-
-val recentlyPlayedMusicModels = listOf(
-    MusicModel(
-        title = "The Forest Fall",
-        subtitle = "R&B",
-        imageUrl = "https://preview.redd.it/6e5lnfxovhs41.png?width=600&format=png&auto=webp&s=a59f248a0eeb1d636b03952eb4469b578ebdab7d",
-    ),
-    MusicModel(
-        title = "The Rejecteds",
-        subtitle = "Pop",
-        imageUrl = "https://preview.redd.it/z4otbf8vwhs41.png?width=600&format=png&auto=webp&s=d4d3e7305ee20b820f80b785f685e4f87bf03807",
-    ),
-    MusicModel(
-        title = "Lucia Melin",
-        subtitle = "Pop",
-        imageUrl = "https://preview.redd.it/v7j19hz3xhs41.png?width=600&format=png&auto=webp&s=9fd58a2816368c3e6c6a63a585fd5789aa9b0d38",
-    )
-)
-
-val justReleasedMusicModels = listOf(
-    MusicModel(
-        title = "Peace of Mind",
-        subtitle = "Soul",
-        imageUrl = "https://www.yardbarker.com/media/1/f/1ffd1e8d9e270adc2d93507f77e452ae1be3c8ce/600_wide/21MBV.jpg",
-    ),
-    MusicModel(
-        title = "Friends",
-        subtitle = "Acoustic Pop",
-        imageUrl = "https://music-artwork.com/wp-content/uploads/2019/08/dark_side_pink_floyd.jpg",
-    ),
-    MusicModel(
-        title = "Memoria",
-        subtitle = "Alternative Indie",
-        imageUrl = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSV6XZ3-hxgm_ads6pGg5dvqFNYtIp8YyOsWQ&usqp=CAU",
-    ),
-)
-
-val justForYouMusicModels = listOf(
-    MusicModel(
-        title = "The Forest Fall",
-        subtitle = "R&B",
-        imageUrl = "https://i.pinimg.com/originals/ae/78/41/ae7841d2c1f2fd60a863d97c5ce2446b.jpg",
-    ),
-    MusicModel(
-        title = "The Rejecteds",
-        subtitle = "Pop",
-        imageUrl = "https://www.billboard.com/wp-content/uploads/2022/03/20.-Joy-Division-%E2%80%98Unknown-Pleasures-1979-album-art-billboard-1240.jpg?w=600",
-    ),
-    MusicModel(
-        title = "Lucia Melin",
-        subtitle = "Pop",
-        imageUrl = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTKYEc3YP0R2wv0jk4SMyOoHosddzzSeApTEw&usqp=CAU",
-    ),
-)
-
-val artistListModels = listOf(
-    ArtistModel(
-        name = "Bruno Mars",
-        songCount = 4,
-        imageUrl = "https://flxt.tmsimg.com/assets/575770_v9_bg.jpg"
-    ),
-    ArtistModel(
-        name = "Ed Sheeran",
-        songCount = 8,
-        imageUrl = "https://ca-times.brightspotcdn.com/dims4/default/e28f00c/2147483647/strip/true/crop/3500x2333+0+0/resize/840x560!/quality/80/?url=https%3A%2F%2Fcalifornia-times-brightspot.s3.amazonaws.com%2F17%2Fe4%2F01486f774df688d37a5d12bbcb69%2Fjazz-fest-26819.jpg"
-    ),
-    ArtistModel(
-        name = "Lady Gaga",
-        songCount = 12,
-        imageUrl = "https://cdn.britannica.com/57/136557-050-0FFC3C93/Lady-Gaga-ceremony-Grammy-Awards-Los-Angeles-Jan-31-2010.jpg"
-    ),
-    ArtistModel(
-        name = "Bad Bunny",
-        songCount = 7,
-        imageUrl = "https://yt3.googleusercontent.com/7tCfeCWH4arhsTM-4Rz4IxWieQbegzibeXlG-kbytAujdk5dr2K0gBb8NG0Cvk6lB1dPkjyd=s900-c-k-c0x00ffffff-no-rj"
-    ),
-    ArtistModel(
-        name = "Rosalia",
-        songCount = 10,
-        imageUrl = "https://i.insider.com/5d4d84ff21214c4fe5514038?width=1136&format=jpeg"
-    ),
-    ArtistModel(
-        name = "Rihanna",
-        songCount = 15,
-        imageUrl = "https://assets.vogue.com/photos/63e67c2653cf7705e1e06f3a/4:3/w_2560%2Cc_limit/GettyImages-1246958447.jpg"
-    ),
-    ArtistModel(
-        name = "Kendrick Lamar",
-        songCount = 4,
-        imageUrl = "https://www.nme.com/wp-content/uploads/2022/08/Kendrick-Lamar-2000-x-1270.jpg"
-    ),
-    ArtistModel(
-        name = "J. Cole",
-        songCount = 8,
-        imageUrl = "https://media.pitchfork.com/photos/609d56da35e34915b3b56668/4:3/w_1700,h_1275,c_limit/J-Cole.jpg"
-    )
-)
