+++ android/app-newm/src/main/java/io/newm/screens/investment/portfolio/InvestmentPortfolioPresenter.kt
@@ -1,15 +1,10 @@
 package io.newm.screens.investment.portfolio
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import com.google.android.recaptcha.RecaptchaAction
 import com.slack.circuit.retained.collectAsRetainedState
 import com.slack.circuit.runtime.Navigator
@@ -21,7 +16,6 @@ import io.newm.shared.public.usecases.HasWalletConnectionsUseCase
 import io.newm.shared.public.usecases.SyncWalletConnectionsUseCase
 import io.newm.shared.public.usecases.WalletNFTTracksUseCase
 import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.launch
 
 class InvestmentPortfolioPresenter(
     private val navigator: Navigator,
