--- android/app-newm/src/main/java/io/newm/screens/library/LibraryData.kt
@@ -1,158 +1,0 @@
-package io.newm.screens.library
-
-data class SongModel(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val imageUrl: String,
-    val streamUrl: String,
-    val duration: Long,
-    val isDownloaded: Boolean
-)
-
-val savedSongModels = listOf(
-    SongModel(
-        id = "1",
-        title = "Pink Dream",
-        artist = "Jules May",
-        imageUrl = "https://i.redd.it/o676ax0heme71.png",
-        streamUrl = "TBD",
-        duration = 160,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "2",
-        title = "Happy Cooking",
-        artist = "Myan Mia",
-        imageUrl = "https://static-cse.canva.com/blob/1078769/1600w-fxYFTKLArdY.jpg",
-        streamUrl = "TBD",
-        duration = 165,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "3",
-        title = "Sunset Vibes",
-        artist = "Zac Forland",
-        imageUrl = "https://static-cse.canva.com/blob/1078773/1600w-1Nr6gsUndKw.jpg",
-        streamUrl = "TBD",
-        duration = 210,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "4",
-        title = "Jogging",
-        artist = "Disali Merry",
-        imageUrl = "https://i.guim.co.uk/img/media/a561b92bd05b3caf63e0a592d2c9d0b25710bad5/0_302_3200_1920/master/3200.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=7c9b5827f44d8134f6433b9a3241fc69",
-        streamUrl = "TBD",
-        duration = 187,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "5",
-        title = "Go all out",
-        artist = "Lopesiada",
-        imageUrl = "https://i.redd.it/o676ax0heme71.png",
-        streamUrl = "TBD",
-        duration = 169,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "6",
-        title = "Relaxing at home",
-        artist = "The Arquitect",
-        imageUrl = "https://hips.hearstapps.com/hmg-prod/images/soothing-decorating-ideas-dobbies-1648154916.jpg",
-        streamUrl = "TBD",
-        duration = 148,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "7",
-        title = "Mindless Work",
-        artist = "Abstract",
-        imageUrl = "https://images.pexels.com/photos/3183150/pexels-photo-3183150.jpeg?cs=srgb&dl=pexels-fauxels-3183150.jpg&fm=jpg",
-        streamUrl = "TBD",
-        duration = 162,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "8",
-        title = "I don't give af",
-        artist = "Mewmouth",
-        imageUrl = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQOf0OWS3dkZael9oZqn7Dx-jii9eLjLzP3Jw&usqp=CAU",
-        streamUrl = "TBD",
-        duration = 174,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "9",
-        title = "Pink Dream",
-        artist = "Jules May",
-        imageUrl = "https://i.redd.it/o676ax0heme71.png",
-        streamUrl = "TBD",
-        duration = 196,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "10",
-        title = "Happy Cooking",
-        artist = "Myan Mia",
-        imageUrl = "https://static-cse.canva.com/blob/1078769/1600w-fxYFTKLArdY.jpg",
-        streamUrl = "TBD",
-        duration = 166,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "11",
-        title = "Sunset Vibes",
-        artist = "Zac Forland",
-        imageUrl = "https://static-cse.canva.com/blob/1078773/1600w-1Nr6gsUndKw.jpg",
-        streamUrl = "TBD",
-        duration = 161,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "12",
-        title = "Jogging",
-        artist = "Disali Merry",
-        imageUrl = "https://i.guim.co.uk/img/media/a561b92bd05b3caf63e0a592d2c9d0b25710bad5/0_302_3200_1920/master/3200.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=7c9b5827f44d8134f6433b9a3241fc69",
-        streamUrl = "TBD",
-        duration = 157,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "13",
-        title = "Go all out",
-        artist = "Lopesiada",
-        imageUrl = "https://i.redd.it/o676ax0heme71.png",
-        streamUrl = "TBD",
-        duration = 196,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "14",
-        title = "Relaxing at home",
-        artist = "The Arquitect",
-        imageUrl = "https://hips.hearstapps.com/hmg-prod/images/soothing-decorating-ideas-dobbies-1648154916.jpg",
-        streamUrl = "TBD",
-        duration = 148,
-        isDownloaded = true
-    ),
-    SongModel(
-        id = "15",
-        title = "Mindless Work",
-        artist = "Abstract",
-        imageUrl = "https://images.pexels.com/photos/3183150/pexels-photo-3183150.jpeg?cs=srgb&dl=pexels-fauxels-3183150.jpg&fm=jpg",
-        streamUrl = "TBD",
-        duration = 195,
-        isDownloaded = false
-    ),
-    SongModel(
-        id = "16",
-        title = "I don't give af",
-        artist = "Mewmouth",
-        imageUrl = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQOf0OWS3dkZael9oZqn7Dx-jii9eLjLzP3Jw&usqp=CAU",
-        streamUrl = "TBD",
-        duration = 206,
-        isDownloaded = true
-    )
-)
