+++ android/app-newm/src/main/java/io/newm/screens/profile/view/ProfileRecordStorePanel.kt
@@ -1,7 +1,5 @@
 package io.newm.screens.profile.view
 
-import android.content.Intent
-import android.net.Uri
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
