+++ android/app-newm/src/main/java/io/newm/screens/recordstore/RecordStorePresenter.kt
@@ -3,22 +3,14 @@ package io.newm.screens.recordstore
 import android.content.Context
 import android.net.ConnectivityManager
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.platform.LocalContext
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
-import io.newm.screens.profile.view.ProfileUiState
 import io.newm.shared.public.analytics.NewmAppEventLogger
 import io.newm.shared.public.analytics.events.AppScreens
-import io.newm.shared.public.usecases.HasWalletConnectionsUseCase
-import io.newm.shared.public.usecases.UserDetailsUseCase
-import kotlinx.coroutines.flow.distinctUntilChanged
 
 class RecordStorePresenter(
     private val navigator: Navigator,
