+++ android/app-newm/src/main/java/io/newm/utils/ForceAppUpdateViewModel.kt
@@ -3,7 +3,6 @@ package io.newm.utils
 import androidx.lifecycle.ViewModel
 import com.google.android.recaptcha.RecaptchaAction
 import io.newm.feature.login.screen.authproviders.RecaptchaClientProvider
-import io.newm.shared.NewmAppLogger
 import io.newm.shared.public.usecases.ForceAppUpdateUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
