+++ android/app-newm/src/main/java/io/newm/utils/NewmImageLoaderFactory.kt
@@ -4,8 +4,8 @@ import coil3.ImageLoader
 import coil3.PlatformContext
 import coil3.SingletonImageLoader
 import coil3.disk.DiskCache
-import coil3.memory.MemoryCache
 import coil3.disk.directory
+import coil3.memory.MemoryCache
 
 class NewmImageLoaderFactory : SingletonImageLoader.Factory {
     override fun newImageLoader(context: PlatformContext): ImageLoader {
