+++ android/core/ui-utils/src/androidMain/kotlin/io/newm/core/ui/webview/FullScreenWebView.kt
@@ -3,7 +3,6 @@ package io.newm.core.ui.webview
 import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
 import android.webkit.CookieManager
 import android.webkit.WebResourceRequest
 import android.webkit.WebView
@@ -74,6 +73,6 @@ fun FullScreenWebView(
 }
 
 fun launchExternalUrl(context: Context, url: String) {
-    val intent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
+    val intent = Intent(Intent.ACTION_VIEW, url.toUri())
     context.startActivity(intent)
 }
