+++ shared/src/commonMain/kotlin/io.newm.shared/internal/implementations/UserSessionUseCaseImpl.kt
@@ -4,7 +4,6 @@ import io.newm.shared.internal.TokenManager
 import io.newm.shared.public.usecases.UserSessionUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.startWith
 import org.koin.core.component.KoinComponent
 
 internal class UserSessionUseCaseImpl(
